Create Clear-button test; Update-status test
